Fixed bug where stationary angular velocity of falling agent would reset on collision.
Bug happened due to velocity being 0 and since there is no air drag/friction then there is no angular velocity applied to agent as it free falls. Therefore, if no input was given during fall the agent would collide with ground with angular velocity of (0, 0, 0) and trigger an end episode which is not what we want to happen.